+++ com.unity.ml-agents/Runtime/Inference/TensorApplier.cs
@@ -81,15 +81,6 @@ namespace Unity.MLAgents.Inference
                 }
             }
             m_Dict[TensorNames.RecurrentOutput] = new MemoryOutputApplier(memories);
-
-            if (modelVersion < (int)BarracudaModelParamLoader.ModelApiVersion.MLAgents2_0)
-            {
-                for (var i = 0; i < model?.memories.Count; i++)
-                {
-                    m_Dict[model.memories[i].output] =
-                        new BarracudaMemoryOutputApplier(model.memories.Count, i, memories);
-                }
-            }
         }
 
         /// <summary>